Added editing on double click.
The editing works as long as you hit enter once you're done. I was having some trouble getting the blur event to take hold. I think the TaskList might need to be broken up to have a component rendering each item.

Also added the remove functionality, however there is a js error that occurs when doing this :/ not sure what causes it, but it works...