Add montage example to labs
The example has been "mopped" (optimized/bundled), which may
be a bit against the spirit of easily compariing source code to,
but I'm not sure how this would be included otherwise.

Full sourcecode is available. And we're always happy to chat on
irc.freenode.net#montage

Routing support is still under review.