Adding a Vanilla CoffeeScript example
I was looking for a simple TodoMVC version of CoffeeScript and was
surprised to see one didn’t yet exist, so thought I’d create and
contribute it back.

The coffee files are pretty much literal one to one mappings of the
original Vanilla js files.

I checked in the compiled .js files too so the example can be run
without having to compile it.