Refactored react-backbone example to break out components into smaller pieces, and delegated responsibility to those smaller components.
Moved save/destroy/cancel functionality into the todo item
components themselves and out of the app.

Moved filtering of todo items into the collection and out of the app.

Created a TodoList component, which contains the items and keeps track
of which one is currently editing.

Broke out footer into individual components that just grab what they
need from the list, and just modify the list themselves

upgraded to react 0.12.1, so that we can return null from render