Update canjs to use npm not bower.
Ran into 2 big issues with canjs, napa came in handy once more.

For 1 canjs does not come with a built dist file. Installing from npm
does not work since they messed up the 2.0.7 release and forgot to
include a bower dep. thus even installing and building is a no go.

The second pice was @passy's localstorage repo has never been published
to npm thus i had to get it from the tag on the repo.

Needless to say this was plenty of fun.