Remove use of deprecated AppRegions.
Marionette has marked `appRegions` as a deprecated feature because it is a
leaky abstraction for the Application to know so much about the DOM and
the structure of your actual page.
Instead of using `appRegions`, Marionette encourages you to setup a
RootLayout whose concern is setting up render targets.

In the near future we are introducing the concept of a "root" to the
application object
https://github.com/marionettejs/backbone.marionette/issues/2279

So that uses will have an idiomatic way of using this pattern. The goal
here is not to conflate the concerns of splitting up render targets
with your application, rather the goal is to give users a pattern to follow when
setting up a single or multiple application objects.

Fixes #1171
Fixes #1184