Debounce the rendering of the application view
If we don't debounce "this.render", the app-view will be rendered 4 times on each todo (visible, change, change:complete and sync). So when we try to check globally all todos, this.render is called a hundred times. Once this PR is accepted, I will edit Backbone-requirejs.

Another problem I have addressed concerning the global checkbox. It should be checked when we have at least one completed todo.