tests: Fix flakey list tests.
This test was causing us issues on CI for several reasons. The first
being that we were touching the DOM outside of webdrivers promise loop
(via Q). The second cause of issues was that we were holding onto DOM
nodes for a long time after we first queried the DOM, allowing the
pointers to potentially go stale.