Drop `derby` app
The app has been down for multiple years, was left unmaintained for
really long and is outdated in many ways. Thus, we're (temporarily)
removing it. If somebody comes in, volunteers to fix the problems and we
find a place to host it.

We already decided earlier to drop it from the homepage.
Ref 7f59ef71a0b55fe70cc8a8c95a3d374a1894deca

Fixes gh-546
Fixes gh-593