Request for adding diffHTML to TodoMVC
Hi TasteJS folks!

This is a PR that contains a snapshot of:
https://github.com/tbranyen/todomvc to be included as part of TodoMVC's
collection.

This utilizes a tool I've been working on called
[diffHTML](http://www.diffhtml.org) that helps write components and
applications with declarative *real* HTML and hook into Virtual DOM
changes with transition hooks.

It also utilizes Redux to handle all state, since it pairs so nicely
with diffHTML it made sense to include over a vanilla approach.

Let me know what you think, thanks for your consideration!