Updates example to get tests passing
This commit sets the node_modules/todomvc-app-css and common to fixed
files found in the backbone example which pass the tests. Don't think
the tests were updated for the latest styles.

Removed inline transitions and transitions in general since they were
messing up the browser tests. It yields a cleaner source as well.

Updated the README