Updating "Why I Like TypeScript" screencast link
Existing link for the screencast was leading to [a rather empty looking page](http://www.leebrimelow.com/why-i-like-typescripts).

I updated the link to what I believe is the same screencast the original link was referring to. The link redirects to the video "Why I Like TypeScript" on YouTube by Lee Brimelow.